Improve our performance demo.
There is still a fair amount of work to be done before we can be super
confident in the results here, but this feels like a good step forward:

* Tests all run in an iframe.
* Progress bar is printed to improve feedback.
* Long text moved to “PERFORMANCE.md”.
