+++ demo/performance/lit-html.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <script type="importmap">
+      {
+        "imports": {
+          "lit-html": "https://esm.sh/lit-html@3.2.1"
+        }
+      }
+    </script>
+    <link rel="stylesheet" href="./index.css">
+  </head>
+  <body>
+    <script type="module" src="./lit-html.js"></script>
+  </body>
+</html>
