+++ demo/performance/react.html
@@ -1,0 +1,18 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <script type="importmap">
+      {
+        "imports": {
+          "react": "https://esm.sh/react@18.3.1",
+          "react-dom/": "https://esm.sh/react-dom@18.3.1/"
+        }
+      }
+    </script>
+    <link rel="stylesheet" href="./index.css">
+  </head>
+  <body>
+    <script type="module" src="./react.js"></script>
+  </body>
+</html>
