+++ demo/performance/uhtml.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <script type="importmap">
+      {
+        "imports": {
+          "uhtml": "https://esm.sh/uhtml@4.5.11"
+        }
+      }
+    </script>
+    <link rel="stylesheet" href="./index.css">
+  </head>
+  <body>
+    <script type="module" src="./uhtml.js"></script>
+  </body>
+</html>
